+++ crates/wasm-mutate/src/mutators.rs
@@ -105,7 +105,7 @@ pub trait Mutator {
 }
 
 /// Type helper to wrap operator and the byte offset in the code section of a Wasm module
-pub type OperatorAndByteOffset<'a> = (Operator<'a>, usize);
+pub type OperatorAndByteOffset<'a> = (Operator<'a>, u64);
 
 #[cfg(test)]
 fn match_mutation<T>(original: &str, mutator: T, expected: &str)
